refactor codes in EstimatorHelper and move TuplespaceEvent to rspace package
