+++ rspace/src/main/scala/coop/rchain/rspace/trace/Event.scala
@@ -5,12 +5,13 @@ import coop.rchain.rspace.StableHashProvider._
 import coop.rchain.rspace.internal._
 import cats.implicits._
 import coop.rchain.rspace.Blake2b256Hash
-import coop.rchain.shared.Serialize
 import coop.rchain.rspace.internal.codecSeq
+import coop.rchain.rspace.trace.TuplespaceEvent.TuplespaceEventOps
 import coop.rchain.shared.Serialize
 import scodec.Codec
 import scodec.bits.ByteVector
 import scodec.codecs._
+import shapeless.syntax.std.tuple.productTupleOps
 
 import scala.collection.SortedSet
 
@@ -38,6 +39,145 @@ object Event {
       }(Codec[Consume])
 
   implicit def codecLog: Codec[Seq[Event]] = codecSeq[Event](codecEvent)
+
+  /**
+    * `Scope` is widely used in descriptions in this file. To better understand `scope` here. I'd like to present
+    * some examples to demonstrate that.
+    *
+    * B2  Rho("for (_ <<- @1) { 0 }")
+    * |
+    * B1  Rho("@1!(0)")
+    *
+    * I'd like to use `Volatile` event here to demonstrate the `scope` here.
+    *
+    * Within the scope which Scope(B2, B3)
+    *
+    * B2  Rho("for (_ <<- @1) { 0 }") ----+
+    * |                                   |---Scope(B2, B1)
+    * B1  Rho("@1!(0)")               ----+
+    *
+    * If you look at within the scope of Scope(B2, B1) like above, all the comm events happened in B1 and B2 are Volatile.
+    *
+    *
+    * B2  Rho("for (_ <<- @1) { 0 }") ---- Scope(B2)
+    * |
+    * B1  Rho("@1!(0)")
+    *
+    * But if you look at within only the scope os Scope(B2) like above, the comm and consume events in the B2
+    * are not Volatile.
+    *
+    * `Volatile` means that the produce event and the consume event match within the scope.
+    * This produce and consume events within this comm event would not leave any thing left in the
+    * tuplespace within the scope.
+    */
+  def isVolatile(comm: COMM, consumes: Set[Consume], produces: Set[Produce]) =
+    !comm.consume.persistent && comm.peeks.isEmpty && consumes.contains(comm.consume) && comm.produces
+      .forall(
+        produce => !produce.persistent && produces.contains(produce)
+      )
+
+  /**
+    * @return EventGroup contains all the events which is not volatile in the scope
+    */
+  def extractRSpaceEventGroup(events: List[Event]): EventGroup = {
+    val rspaceEvents = events.toSet
+    val (allProduceEvents, allConsumeEvents, allCommEvents) =
+      rspaceEvents.foldLeft((Seq.empty[Produce], Seq.empty[Consume], Seq.empty[COMM])) {
+        (op, event) =>
+          event match {
+            case p: Produce => (op._1 :+ p, op._2, op._3)
+            case c: Consume => (op._1, op._2 :+ c, op._3)
+            case c: COMM    => (op._1, op._2, op._3 :+ c)
+          }
+      }
+    val (volatileCommEvents, nonVolatileCommEvents) = allCommEvents
+      .partition(isVolatile(_, allConsumeEvents.toSet, allProduceEvents.toSet))
+    val producesInVolatileCommEvents = volatileCommEvents.flatMap(_.produces)
+    val consumesInVolatileCommEvents = volatileCommEvents.map(_.consume)
+    val produceEvents                = allProduceEvents.filterNot(producesInVolatileCommEvents.contains)
+    val consumeEvents                = allConsumeEvents.filterNot(consumesInVolatileCommEvents.contains)
+    EventGroup(produceEvents.toSet, consumeEvents.toSet, nonVolatileCommEvents.toSet)
+  }
+
+  def extractJoinedChannels(b: EventGroup): Set[Blake2b256Hash] = {
+
+    def joinedChannels(consumes: Set[Consume]) =
+      consumes.withFilter(Consume.hasJoins).flatMap(_.channelsHashes)
+
+    joinedChannels(b.consumes) ++ joinedChannels(b.comms.map(_.consume))
+
+  }
+  def produceChannels(events: EventGroup) =
+    events.produces.map(_.channelsHash).toSet ++ events.comms.flatMap { comm =>
+      comm.produces.map(_.channelsHash)
+    }.toSet
+
+  def allChannels(events: EventGroup) =
+    events.produces.map(_.channelsHash).toSet ++ events.consumes
+      .flatMap(_.channelsHashes)
+      .toSet ++ events.comms.flatMap { comm =>
+      comm.consume.channelsHashes ++ comm.produces.map(_.channelsHash)
+    }.toSet
+
+  def tuplespaceEventsPerChannel(
+      b: EventGroup
+  ): Map[Blake2b256Hash, Set[TuplespaceEvent]] = {
+    val nonPersistentProducesInComms = b.comms.flatMap(_.produces).filterNot(_.persistent)
+    val nonPersistentConsumesInComms = b.comms.map(_.consume).filterNot(_.persistent)
+    val peekedProducesInCommsHashes =
+      b.comms.withFilter(_.peeks.nonEmpty).flatMap(_.produces).map(_.hash)
+
+    val freeProduces = b.produces
+      .diff(nonPersistentProducesInComms)
+      .filterNot(p => peekedProducesInCommsHashes.contains(p.hash) && !p.persistent)
+
+    val freeConsumes = b.consumes.diff(nonPersistentConsumesInComms)
+
+    val produceEvents = freeProduces.map(TuplespaceEvent.from(_))
+    val consumeEvents = freeConsumes.flatMap(TuplespaceEvent.from(_))
+    val commEvents    = b.comms.flatMap(TuplespaceEvent.from(_, b.consumes))
+
+    (produceEvents ++ consumeEvents ++ commEvents)
+      .groupBy(_._1)
+      .mapValues[Set[TuplespaceEvent]](_.map(_._2))
+  }
+
+  def containConflictingEvents(
+      b1Events: EventGroup,
+      b2Events: EventGroup
+  ): Set[Blake2b256Hash] = {
+    def channelConflicts(
+        b1Events: Set[TuplespaceEvent],
+        b2Events: Set[TuplespaceEvent]
+    ): Boolean =
+      (for {
+        b1  <- b1Events
+        b2  <- b2Events
+        res = b1.conflicts(b2)
+        // TODO: fail fast
+      } yield (res)).contains(true)
+
+    val b1Ops = tuplespaceEventsPerChannel(b1Events)
+    val b2Ops = tuplespaceEventsPerChannel(b2Events)
+    val conflictPerChannel = b1Ops
+      .map {
+        case (channel, v) =>
+          (channel, channelConflicts(v, b2Ops.getOrElse(channel, Set.empty)))
+      }
+    conflictPerChannel
+      .filter { case (_, conflicts) => conflicts }
+      .keys
+      .toSet
+  }
+
+  trait IsConflict
+  final case class NonConflict(leftEvents: EventGroup, rightEvents: EventGroup) extends IsConflict
+  final case class Conflict(
+      leftEvents: EventGroup,
+      rightEvents: EventGroup,
+      conflictSet: Set[Blake2b256Hash]
+  ) extends IsConflict
+
 }
 
 final case class COMM(
