+++ rspace/src/main/scala/coop/rchain/rspace/trace/TuplespaceEvent.scala
@@ -1,7 +1,6 @@
-package coop.rchain.casper
+package coop.rchain.rspace.trace
 
 import coop.rchain.rspace.Blake2b256Hash
-import coop.rchain.rspace.trace._
 
 final case class TuplespaceEvent(
     incoming: TuplespaceOperation,
@@ -67,7 +66,7 @@ object TuplespaceEvent {
 
   implicit class TuplespaceEventOps(val ev: TuplespaceEvent) extends AnyVal {
 
-    private[casper] def conflicts(other: TuplespaceEvent): Boolean =
+    def conflicts(other: TuplespaceEvent): Boolean =
       if (ev.incoming.polarity == other.incoming.polarity) {
 
         val bothPeeks = (ev.incoming.cardinality == Peek) && (other.incoming.cardinality ==
@@ -83,7 +82,7 @@ object TuplespaceEvent {
 
       } else ev.unsatisfied && other.unsatisfied
 
-    private[casper] def unsatisfied: Boolean =
+    def unsatisfied: Boolean =
       ev.incoming.cardinality match {
         case Peek      => ev.matched.isEmpty
         case Linear    => ev.matched.forall(_.cardinality == Peek)
